+++ lib/execution_model.ml
@@ -37,7 +37,7 @@ let completed_bar_v1_contract =
     version = "2";
     previous_versions = [ "1" ];
     scenario_contract_versions =
-      [ "11"; "10"; "9"; "8"; "7"; "6"; "5"; "4"; "3" ];
+      [ "12"; "11"; "10"; "9"; "8"; "7"; "6"; "5"; "4"; "3" ];
     required_fields = [ "version"; "participation_bps"; "fee_schedules" ];
     legacy_required_fields =
       [ "version"; "participation_bps"; "fixed_fee"; "fee_bps" ];